Initial markup and styles for result
Styles are cover the mobile view and not all of them have been modified
to be responsive. Since in result same buttons as for playing is used
the styles for button in result extend the ones used in game__body with
some modifications.

Styels for main__content-wrapper is now seperad in its own partial.