Initial styles for game__result--intro modifier
This modifier changes the behavior of game__result.
Hover state for game__button-house is removed.
Layout of game__result is modified because in this state modifier the
replay button is not visible.
The ::before pseudo element game__house-bttuon is removed
game__house-button is restyles to a cicle contains the sprite.
A new pseudo element contains the sprite for each of --circle- modifiers