Completed the desktop view of modal
Until the first breakpoint 43em modal take the whole screen and it's
positioned relative to the body element. On desktop view it turn into a
box with finite dimensions and it's positioned relative to
game__content-wrapper because this way is way it maintains a consistent
position regardless of viewport height.
A state modifiers was also added whose sole purpose is animation.